fix: revert "fix: jsonGet interpolation issue 2426 + allow step depends on pass in step instance (#2477)"
This reverts commit 4c0d3cf0735a4c9c120a3c75057f8725db493393.
