+++ doc/workflows/pipelines/sagemaker.workflow.pipelines.rst
@@ -6,6 +6,7 @@ ConditionStep
 
 .. autoclass:: sagemaker.workflow.condition_step.ConditionStep
 
+.. autoclass:: sagemaker.workflow.condition_step.JsonGet
 
 Conditions
 ----------
@@ -54,7 +55,6 @@ Functions
 ---------
 
 .. autoclass:: sagemaker.workflow.functions.Join
-.. autoclass:: sagemaker.workflow.functions.JsonGet
 
 Parameters
 ----------
