+++ src/sagemaker/workflow/condition_step.py
@@ -15,9 +15,17 @@ from __future__ import absolute_import
 
 from typing import List, Union
 
+import attr
+
 from sagemaker.workflow.conditions import Condition
-from sagemaker.workflow.entities import RequestType
-from sagemaker.workflow.properties import Properties
+from sagemaker.workflow.entities import (
+    Expression,
+    RequestType,
+)
+from sagemaker.workflow.properties import (
+    Properties,
+    PropertyFile,
+)
 from sagemaker.workflow.steps import (
     Step,
     StepTypeEnum,
@@ -32,7 +40,7 @@ class ConditionStep(Step):
     def __init__(
         self,
         name: str,
-        depends_on: Union[List[str], List[Step]] = None,
+        depends_on: List[str] = None,
         conditions: List[Condition] = None,
         if_steps: List[Union[Step, StepCollection]] = None,
         else_steps: List[Union[Step, StepCollection]] = None,
@@ -76,3 +84,33 @@ class ConditionStep(Step):
     def properties(self):
         """A simple Properties object with `Outcome` as the only property"""
         return self._properties
+
+
+@attr.s
+class JsonGet(Expression):
+    """Get JSON properties from PropertyFiles.
+
+    Attributes:
+        step (Step): The step from which to get the property file.
+        property_file (Union[PropertyFile, str]): Either a PropertyFile instance
+            or the name of a property file.
+        json_path (str): The JSON path expression to the requested value.
+    """
+
+    step: Step = attr.ib()
+    property_file: Union[PropertyFile, str] = attr.ib()
+    json_path: str = attr.ib()
+
+    @property
+    def expr(self):
+        """The expression dict for a `JsonGet` function."""
+        if isinstance(self.property_file, PropertyFile):
+            name = self.property_file.name
+        else:
+            name = self.property_file
+        return {
+            "Std:JsonGet": {
+                "PropertyFile": {"Get": f"Steps.{self.step.name}.PropertyFiles.{name}"},
+                "Path": self.json_path,
+            }
+        }
