+++ src/sagemaker/workflow/functions.py
@@ -13,12 +13,11 @@
 """The step definitions for workflow."""
 from __future__ import absolute_import
 
-from typing import List, Union
+from typing import List
 
 import attr
 
 from sagemaker.workflow.entities import Expression
-from sagemaker.workflow.properties import PropertyFile
 
 
 @attr.s
@@ -44,39 +43,4 @@ class Join(Expression):
                     value.expr if hasattr(value, "expr") else value for value in self.values
                 ],
             },
-        }
-
-
-@attr.s
-class JsonGet(Expression):
-    """Get JSON properties from PropertyFiles.
-
-    Attributes:
-        processing_step_name (str): The step name of the `sagemaker.workflow.steps.ProcessingStep`
-            from which to get the property file.
-        property_file (Union[PropertyFile, str]): Either a PropertyFile instance
-            or the name of a property file.
-        json_path (str): The JSON path expression to the requested value.
-    """
-
-    processing_step_name: str = attr.ib()
-    property_file: Union[PropertyFile, str] = attr.ib()
-    json_path: str = attr.ib()
-
-    @property
-    def expr(self):
-        """The expression dict for a `JsonGet` function."""
-        if isinstance(self.property_file, PropertyFile):
-            name = self.property_file.name
-        else:
-            name = self.property_file
-
-        if not isinstance(self.processing_step_name, str):
-            raise ValueError("processing_step_name passed in is not instance of a str")
-
-        return {
-            "Std:JsonGet": {
-                "PropertyFile": {"Get": f"Steps.{self.processing_step_name}.PropertyFiles.{name}"},
-                "Path": self.json_path,
-            }
         }
