+++ src/sagemaker/workflow/steps.py
@@ -60,13 +60,12 @@ class Step(Entity):
     Attributes:
         name (str): The name of the step.
         step_type (StepTypeEnum): The type of the step.
-        depends_on (List[str] or List[Step]): The list of step names or step
-            instances the current step depends on
+        depends_on (List[str]): The list of step names the current step depends on
     """
 
     name: str = attr.ib(factory=str)
     step_type: StepTypeEnum = attr.ib(factory=StepTypeEnum.factory)
-    depends_on: Union[List[str], List["Step"]] = attr.ib(default=None)
+    depends_on: List[str] = attr.ib(default=None)
 
     @property
     @abc.abstractmethod
@@ -86,13 +85,11 @@ class Step(Entity):
             "Arguments": self.arguments,
         }
         if self.depends_on:
-            request_dict["DependsOn"] = self._resolve_depends_on(self.depends_on)
-
+            request_dict["DependsOn"] = self.depends_on
         return request_dict
 
-    def add_depends_on(self, step_names: Union[List[str], List["Step"]]):
-        """Add step names or step instances to the current step depends on list"""
-
+    def add_depends_on(self, step_names: List[str]):
+        """Add step names to the current step depends on list"""
         if not step_names:
             return
         if not self.depends_on:
@@ -103,19 +100,6 @@ class Step(Entity):
     def ref(self) -> Dict[str, str]:
         """Gets a reference dict for steps"""
         return {"Name": self.name}
-
-    @staticmethod
-    def _resolve_depends_on(depends_on_list: Union[List[str], List["Step"]]):
-        """Resolver the step depends on list"""
-        depends_on = []
-        for step in depends_on_list:
-            if isinstance(step, Step):
-                depends_on.append(step.name)
-            elif isinstance(step, str):
-                depends_on.append(step)
-            else:
-                raise ValueError(f"Invalid input step name: {step}")
-        return depends_on
 
 
 @attr.s
@@ -159,7 +143,7 @@ class TrainingStep(Step):
         estimator: EstimatorBase,
         inputs: Union[TrainingInput, dict, str, FileSystemInput] = None,
         cache_config: CacheConfig = None,
-        depends_on: Union[List[str], List[Step]] = None,
+        depends_on: List[str] = None,
     ):
         """Construct a TrainingStep, given an `EstimatorBase` instance.
 
@@ -187,8 +171,8 @@ class TrainingStep(Step):
                     the path to the training dataset.
 
             cache_config (CacheConfig):  A `sagemaker.workflow.steps.CacheConfig` instance.
-            depends_on (List[str] or List[Step]): A list of step names or step instances
-                this `sagemaker.workflow.steps.TrainingStep` depends on
+            depends_on (List[str]): A list of step names this `sagemaker.workflow.steps.TrainingStep`
+                depends on
         """
         super(TrainingStep, self).__init__(name, StepTypeEnum.TRAINING, depends_on)
         self.estimator = estimator
@@ -233,11 +217,7 @@ class CreateModelStep(Step):
     """CreateModel step for workflow."""
 
     def __init__(
-        self,
-        name: str,
-        model: Model,
-        inputs: CreateModelInput,
-        depends_on: Union[List[str], List[Step]] = None,
+        self, name: str, model: Model, inputs: CreateModelInput, depends_on: List[str] = None
     ):
         """Construct a CreateModelStep, given an `sagemaker.model.Model` instance.
 
@@ -249,8 +229,8 @@ class CreateModelStep(Step):
             model (Model): A `sagemaker.model.Model` instance.
             inputs (CreateModelInput): A `sagemaker.inputs.CreateModelInput` instance.
                 Defaults to `None`.
-            depends_on (List[str] or List[Step]): A list of step names or step instances
-                this `sagemaker.workflow.steps.CreateModelStep` depends on
+            depends_on (List[str]): A list of step names this `sagemaker.workflow.steps.CreateModelStep`
+                depends on
         """
         super(CreateModelStep, self).__init__(name, StepTypeEnum.CREATE_MODEL, depends_on)
         self.model = model
@@ -295,7 +275,7 @@ class TransformStep(Step):
         transformer: Transformer,
         inputs: TransformInput,
         cache_config: CacheConfig = None,
-        depends_on: Union[List[str], List[Step]] = None,
+        depends_on: List[str] = None,
     ):
         """Constructs a TransformStep, given an `Transformer` instance.
 
@@ -307,8 +287,8 @@ class TransformStep(Step):
             transformer (Transformer): A `sagemaker.transformer.Transformer` instance.
             inputs (TransformInput): A `sagemaker.inputs.TransformInput` instance.
             cache_config (CacheConfig):  A `sagemaker.workflow.steps.CacheConfig` instance.
-            depends_on (List[str] or List[Step]): A list of step names or step instances
-                this `sagemaker.workflow.steps.TransformStep` depends on
+            depends_on (List[str]): A list of step names this `sagemaker.workflow.steps.TransformStep`
+                depends on
         """
         super(TransformStep, self).__init__(name, StepTypeEnum.TRANSFORM, depends_on)
         self.transformer = transformer
@@ -371,7 +351,7 @@ class ProcessingStep(Step):
         code: str = None,
         property_files: List[PropertyFile] = None,
         cache_config: CacheConfig = None,
-        depends_on: Union[List[str], List[Step]] = None,
+        depends_on: List[str] = None,
     ):
         """Construct a ProcessingStep, given a `Processor` instance.
 
@@ -392,8 +372,8 @@ class ProcessingStep(Step):
             property_files (List[PropertyFile]): A list of property files that workflow looks
                 for and resolves from the configured processing output list.
             cache_config (CacheConfig):  A `sagemaker.workflow.steps.CacheConfig` instance.
-            depends_on (List[str] or List[Step]): A list of step names or step instance
-                this `sagemaker.workflow.steps.ProcessingStep` depends on
+            depends_on (List[str]): A list of step names this `sagemaker.workflow.steps.ProcessingStep`
+                depends on
         """
         super(ProcessingStep, self).__init__(name, StepTypeEnum.PROCESSING, depends_on)
         self.processor = processor
